Minor Modifications

+++ src/Wallet.sol
@@ -1,31 +1,32 @@
 pragma solidity 0.8.13;
 
-
 contract Wallet {
-    
     address payable public owner;
-    uint public balance;
+    uint256 public balance;
 
-    event Deposit(address account, uint amount);
-    constructor() payable{
+    event Deposit(address account, uint256 amount);
+
+    constructor() payable {
         owner = payable(msg.sender);
     }
 
-    function withdrawEther(uint amount) external{
+    function withdrawEther(uint256 amount) external {
         require(msg.sender == owner, "Only owner can call this function");
         require(amount <= balance, "Amount exceeds balance");
         balance -= amount;
         payable(owner).transfer(amount);
     }
 
-    function setNewOwner(address _owner) external{
+    function setNewOwner(address _owner) external {
         require(msg.sender == owner, "Only owner can call this function");
         owner = payable(_owner);
     }
-    receive() external payable{
+
+    receive() external payable {
         emit Deposit(msg.sender, msg.value);
     }
-    fallback() external payable{
+
+    fallback() external payable {
         emit Deposit(msg.sender, msg.value);
     }
 }
